@@ -1,11 +1,12 @@
 package github.yaa110.memento.adapter;
 
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.ArrayList;
 
@@ -37,12 +38,7 @@ public class FolderAdapter extends RecyclerView.Adapter<FolderAdapter.ViewHolder
 			holder.icon.setImageResource(R.drawable.ic_file);
 		}
 
-		holder.holder.setOnClickListener(new View.OnClickListener() {
-			@Override
-			public void onClick(View view) {
-				listener.onClick(item);
-			}
-		});
+		holder.holder.setOnClickListener(view -> listener.onClick(item));
 	}
 
 	@Override
@@ -58,8 +54,8 @@ public class FolderAdapter extends RecyclerView.Adapter<FolderAdapter.ViewHolder
 		public ViewHolder(View itemView) {
 			super(itemView);
 			holder = itemView.findViewById(R.id.holder);
-			title = (TextView) itemView.findViewById(R.id.title);
-			icon = (ImageView) itemView.findViewById(R.id.icon);
+			title = itemView.findViewById(R.id.title);
+			icon = itemView.findViewById(R.id.icon);
 		}
 	}
 
