@@ -5,16 +5,17 @@ import android.os.Bundle;
 import android.os.Environment;
 import android.os.Handler;
 import android.os.Looper;
-import android.support.annotation.Nullable;
-import android.support.annotation.StringRes;
-import android.support.v4.app.DialogFragment;
-import android.support.v7.widget.LinearLayoutManager;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.Window;
 import android.widget.EditText;
 import android.widget.TextView;
+
+import androidx.annotation.Nullable;
+import androidx.annotation.StringRes;
+import androidx.fragment.app.DialogFragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
 
 import java.io.File;
 import java.io.FileFilter;
@@ -90,12 +91,7 @@ public class SaveDialog extends DialogFragment {
 					} catch (Exception ignored) {
 						listener.onError();
 					} finally {
-						new Handler(Looper.getMainLooper()).post(new Runnable() {
-							@Override
-							public void run() {
-								dismiss();
-							}
-						});
+						new Handler(Looper.getMainLooper()).post(() -> dismiss());
 						interrupt();
 					}
 				}
@@ -104,119 +100,104 @@ public class SaveDialog extends DialogFragment {
 			((TextView) view.findViewById(R.id.title_txt)).setText(getString(title));
 			current_path = App.last_path != null ? App.last_path : Environment.getExternalStorageDirectory().getAbsolutePath();
 
-			recyclerView = (FixedHeightRecyclerView) view.findViewById(R.id.recyclerView);
+			recyclerView = view.findViewById(R.id.recyclerView);
 			items = new ArrayList<>();
 			reload();
 
-			view.findViewById(R.id.positive_btn).setOnClickListener(new View.OnClickListener() {
-				@SuppressWarnings("ResultOfMethodCallIgnored")
-				@Override
-				public void onClick(View view) {
-					if (isWorking) return;
-					isWorking = true;
-					new Thread() {
-						@SuppressWarnings("ConstantConditions")
-						@Override
-						public void run() {
+			view.findViewById(R.id.positive_btn).setOnClickListener(view1 -> {
+				if (isWorking) return;
+				isWorking = true;
+				new Thread() {
+					@SuppressWarnings("ConstantConditions")
+					@Override
+					public void run() {
+						try {
 							try {
+								if (isDirWritable()) {
+									App.last_path = current_path;
+									App.instance.putPrefs(App.LAST_PATH_KEY, current_path);
+									pathSelected();
+								}
+							} catch (Exception ignored) {
 								try {
-									if (isDirWritable()) {
-										App.last_path = current_path;
-										App.instance.putPrefs(App.LAST_PATH_KEY, current_path);
-										pathSelected();
-									}
-								} catch (Exception ignored) {
+									current_path = getContext().getExternalFilesDir(null).getAbsolutePath();
+									if (isDirWritable()) pathSelected();
+								} catch (Exception ignored2) {
+									current_path = getContext().getFilesDir().getAbsolutePath();
 									try {
-										current_path = getContext().getExternalFilesDir(null).getAbsolutePath();
 										if (isDirWritable()) pathSelected();
-									} catch (Exception ignored2) {
-										current_path = getContext().getFilesDir().getAbsolutePath();
-										try {
-											if (isDirWritable()) pathSelected();
-										} catch (Exception e) {
-											listener.onError();
-										}
+									} catch (Exception e) {
+										listener.onError();
 									}
 								}
-							} finally {
-								new Handler(Looper.getMainLooper()).post(new Runnable() {
-									@Override
-									public void run() {
-										canceled = false;
-										dismiss();
-									}
-								});
-								interrupt();
 							}
+						} finally {
+							new Handler(Looper.getMainLooper()).post(() -> {
+								canceled = false;
+								dismiss();
+							});
+							interrupt();
 						}
-					}.start();
-				}
+					}
+				}.start();
 			});
 
-			view.findViewById(R.id.negative_btn).setOnClickListener(new View.OnClickListener() {
-				@Override
-				public void onClick(View view) {
-					if (isWorking) return;
-					canceled = true;
-					dismiss();
-				}
+			view.findViewById(R.id.negative_btn).setOnClickListener(view2 -> {
+				if (isWorking) return;
+				canceled = true;
+				dismiss();
 			});
 
-			view.findViewById(R.id.new_btn).setOnClickListener(new View.OnClickListener() {
-				@Override
-				public void onClick(View view) {
-					ContentDialog.newInstance(
-						R.string.new_folder,
-						R.string.create,
-						R.string.cancel,
-						-1,
-						R.layout.dialog_new_folder,
-						new ContentDialog.DialogListener() {
-							private EditText name_txt;
-							private boolean isCreating = false;
+			view.findViewById(R.id.new_btn).setOnClickListener(view3 -> ContentDialog.newInstance(
+				R.string.new_folder,
+				R.string.create,
+				R.string.cancel,
+				-1,
+				R.layout.dialog_new_folder,
+				new ContentDialog.DialogListener() {
+					private EditText name_txt;
+					private boolean isCreating = false;
 
-							@Override
-							public void onPositive(ContentDialog dialog, View content) {
-								if (isCreating || !dialog.checkEditText(name_txt)) return;
-								isCreating = true;
+					@Override
+					public void onPositive(ContentDialog dialog, View content) {
+						if (isCreating || !dialog.checkEditText(name_txt)) return;
+						isCreating = true;
 
-								try {
-									String folder_name = name_txt.getText().toString();
-									File folder = new File(current_path, folder_name);
+						try {
+							String folder_name = name_txt.getText().toString();
+							File folder = new File(current_path, folder_name);
 
-									int counter = 2;
-									while (folder.exists()) {
-										folder = new File(current_path, String.format(Locale.US, "%s(%d)", folder_name, counter));
-										counter++;
-									}
-
-									//noinspection ResultOfMethodCallIgnored
-									folder.mkdirs();
-									reload();
-								} catch (Exception ignored) {
-								} finally {
-									dialog.dismiss();
-								}
+							int counter = 2;
+							while (folder.exists()) {
+								folder = new File(current_path, String.format(Locale.US, "%s(%d)", folder_name, counter));
+								counter++;
 							}
 
-							@Override
-							public void onNegative(ContentDialog dialog, View content) {
-								if (isCreating) return;
-								dialog.dismiss();
-							}
-
-							@Override
-							public void onNeutral(ContentDialog dialog, View content) {
-							}
-
-							@Override
-							public void onInit(View content) {
-								name_txt = (EditText) content.findViewById(R.id.name_txt);
-							}
+							//noinspection ResultOfMethodCallIgnored
+							folder.mkdirs();
+							reload();
+						} catch (Exception ignored) {
+						} finally {
+							dialog.dismiss();
 						}
-					).show(getFragmentManager(), "");
+					}
+
+					@Override
+					public void onNegative(ContentDialog dialog, View content) {
+						if (isCreating) return;
+						dialog.dismiss();
+					}
+
+					@Override
+					public void onNeutral(ContentDialog dialog, View content) {
+					}
+
+					@Override
+					public void onInit(View content) {
+						name_txt = (EditText) content.findViewById(R.id.name_txt);
+					}
 				}
-			});
+			).show(getFragmentManager(), ""));
 		}
 	}
 
@@ -244,19 +225,9 @@ public class SaveDialog extends DialogFragment {
 				File folder = new File(current_path);
 				if (!folder.exists()) folder = new File(Environment.getExternalStorageDirectory().getAbsolutePath());
 
-				File[] folders = folder.listFiles(new FileFilter() {
-					@Override
-					public boolean accept(File file) {
-						return file.isDirectory();
-					}
-				});
+				File[] folders = folder.listFiles(file -> file.isDirectory());
 
-				Arrays.sort(folders, new Comparator<File>() {
-					@Override
-					public int compare(File f1, File f2) {
-						return f1.getName().compareToIgnoreCase(f2.getName());
-					}
-				});
+				Arrays.sort(folders, (f1, f2) -> f1.getName().compareToIgnoreCase(f2.getName()));
 
 				if (!folder.getAbsolutePath().equals(Environment.getExternalStorageDirectory().getAbsolutePath())) {
 					File parent = folder.getParentFile();
@@ -267,24 +238,21 @@ public class SaveDialog extends DialogFragment {
 					items.add(new Folder(file.getName(), file.getAbsolutePath(), false));
 				}
 
-				new Handler(Looper.getMainLooper()).post(new Runnable() {
-					@Override
-					public void run() {
-						if (adapter == null) {
-							adapter = new FolderAdapter(items, new FolderAdapter.ClickListener() {
-								@Override
-								public void onClick(Folder item) {
-									if (isWorking) return;
-									current_path = item.path;
-									reload();
-								}
-							});
+				new Handler(Looper.getMainLooper()).post(() -> {
+					if (adapter == null) {
+						adapter = new FolderAdapter(items, new FolderAdapter.ClickListener() {
+							@Override
+							public void onClick(Folder item) {
+								if (isWorking) return;
+								current_path = item.path;
+								reload();
+							}
+						});
 
-							recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
-							recyclerView.setAdapter(adapter);
-						} else {
-							adapter.notifyDataSetChanged();
-						}
+						recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
+						recyclerView.setAdapter(adapter);
+					} else {
+						adapter.notifyDataSetChanged();
 					}
 				});
 
